requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
mailjet-rest>=1.3.0
python-dotenv>=1.0.0
